fix: fixed cohesion snippet loading
